Add 'hold' piece feature
Tapping the CTRL key (left or right) will place the current piece in the
"hold" queue and set the current piece to the next piece. The hold piece
can then later be swapped with the current piece by tapping CTRL again.
This is a feature present in modern Tetrises.

The hold swap can only be executed once per falling piece.