prune back use of language features
Originally this tetris game was an early milestone for the language as a
non-trivial demo. As such, it relied heavily on C importing to reduce the
amount of the language that it exercised, and used fixed sized arrays
instead of memory allocation.

Since then it has grown a few more dependencies on later-stage language
features, but I'm going to prune it back a bit so that we can have a
happy little self-hosted compiler demo. As self-hosted grows more
language support, the improvements can be added back.

Changes made are:

 * Updating to discard unused parameters.
 * Updating to new usingnamespace semantics.
 * Updating to newest unit test syntax.
 * Calling libc printf instead of std.debug.panic.
 * Calling libc malloc instead of std.heap.c_allocator.alloc.
 * Calling libc srand/rand instead of std.crypto/std.rand.
 * Calling libc math functions instead of std.math functions.